mock: Move database functions to mock.
