+++ core/src/mock/database.rs
@@ -3,7 +3,9 @@
 //! This module provides mock database interfaces, for testing.
 
 use super::common;
-use crate::{config::BridgeConfig, database::Database, utils::initialize_logger};
+use crate::{
+    config::BridgeConfig, database::Database, errors::BridgeError, utils::initialize_logger,
+};
 use std::thread;
 
 /// Creates a temporary database for testing.
@@ -21,7 +23,7 @@ pub async fn create_test_config(db_name: &str, config_file: &str) -> BridgeConfi
 
     let mut config = common::get_test_config(config_file).unwrap();
     config.db_name = db_name.to_owned();
-    Database::initialize_database(&config).await.unwrap();
+    initialize_database(&config).await.unwrap();
 
     config
 }
@@ -54,4 +56,119 @@ pub async fn create_test_config_with_thread_name(
         + &suffix;
 
     create_test_config(&handle, config_file).await
+}
+
+/// Initializes a new database with given configuration. If the database is
+/// already initialized, it will be dropped before initialization. Meaning,
+/// a clean state is guaranteed.
+///
+/// [`Database::new`] must be called after this to connect to the
+/// initialized database.
+///
+/// **Warning:** This must not be used in release environments and is only
+/// suitable for testing.
+///
+/// TODO: This function must be marked with `#[cfg(test)]` to prevent it
+/// from infiltrating the binaries. See:
+/// https://github.com/chainwayxyz/clementine/issues/181
+pub async fn initialize_database(config: &BridgeConfig) -> Result<(), BridgeError> {
+    drop_database(config).await?;
+
+    create_database(config).await?;
+
+    Database::run_schema_script(config).await?;
+
+    Ok(())
+}
+
+/// Creates a new database with given configuration.
+///
+/// # Errors
+///
+/// Will return [`BridgeError`] if there was a problem with database
+/// connection.
+async fn create_database(config: &BridgeConfig) -> Result<(), BridgeError> {
+    let url = Database::get_postgresql_url(config);
+    let conn = sqlx::PgPool::connect(url.as_str()).await?;
+
+    sqlx::query(&format!(
+        "CREATE DATABASE {} WITH OWNER {}",
+        config.db_name, config.db_user
+    ))
+    .execute(&conn)
+    .await?;
+
+    conn.close().await;
+    Ok(())
+}
+
+/// Drops the database for the given configuration, if it exists.
+///
+/// # Errors
+///
+/// Will return [`BridgeError`] if there was a problem with database
+/// connection. It won't return any errors if the database does not already
+/// exist.
+async fn drop_database(config: &BridgeConfig) -> Result<(), BridgeError> {
+    let url = Database::get_postgresql_url(config);
+    let conn = sqlx::PgPool::connect(url.as_str()).await?;
+
+    let query = format!("DROP DATABASE IF EXISTS {}", &config.db_name);
+    sqlx::query(&query).execute(&conn).await?;
+
+    conn.close().await;
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::{
+        database::Database,
+        mock::{
+            common::get_test_config,
+            database::{create_database, drop_database},
+        },
+    };
+
+    #[tokio::test]
+    async fn create_drop_database() {
+        let mut config = get_test_config("test_config.toml").unwrap();
+        config.db_name = "create_drop_database".to_string();
+
+        // Drop database (clear previous test run artifacts) and check that
+        // connection can't be established.
+        drop_database(&config).await.unwrap();
+        assert!(Database::new(&config).await.is_err());
+
+        // It should be possible to connect new database after creating it.
+        create_database(&config).await.unwrap();
+        Database::new(&config).await.unwrap();
+
+        // Dropping database again should result connection to not be
+        // established.
+        drop_database(&config).await.unwrap();
+        assert!(Database::new(&config).await.is_err());
+    }
+
+    #[tokio::test]
+    async fn initialize_database() {
+        let mut config = get_test_config("test_config.toml").unwrap();
+        config.db_name = "initialize_database".to_string();
+
+        // Drop database (clear previous test run artifacts) and check that
+        // connection can't be established.
+        drop_database(&config).await.unwrap();
+        assert!(Database::new(&config).await.is_err());
+
+        // It should be possible to initialize and connect to the new database.
+        crate::mock::database::initialize_database(&config)
+            .await
+            .unwrap();
+        Database::new(&config).await.unwrap();
+
+        // Dropping database again should result connection to not be
+        // established.
+        drop_database(&config).await.unwrap();
+        assert!(Database::new(&config).await.is_err());
+    }
 }
